Programmer: Michael Vuong Group: Recreation Outlet (Right Minded) Description: RecRPC is now displayed instead of 0 when scanning an item with the barcode scanner (For reals this time)

+++ Retail/RecreationOutletPOS/RecreationOutletPOS/Combined.cs
@@ -666,11 +666,11 @@ namespace RecreationOutletPOS
                     {
                         foreach (DataRow row in ds.Tables["Results"].Rows)
                         {
-                            int id;
+                            long id;
                             string name;
                             double price;
 
-                            int.TryParse(row["RecRPC"].ToString(), out id);
+                            long.TryParse(row["RecRPC"].ToString(), out id);
                             name = (row["Name"].ToString());
                             double.TryParse(row["SellPrice"].ToString(), out price);
